\name{DBDA2-package}
\alias{DBDA2-package}
\alias{DBDA2}
\docType{package}
\title{
\packageTitle{DBDA2}
}
\description{
\packageDescription{DBDA2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DBDA2}
\packageIndices{DBDA2}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{DBDA2}

Maintainer: \packageMaintainer{DBDA2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
